+++ examples/mps/matrix-multiplication/main.rs
@@ -5,6 +5,7 @@ use std::ops::{AddAssign, Mul};
 
 use rand::{thread_rng, Rng};
 
+use metal::mps::matrix::*;
 use metal::mps::*;
 use metal::*;
